odoo_installer/requirements.txt
python-time
python-nginx
requests
urllib3

